fix bug with anonymus and user shopcart

+++ routes/cart.js
@@ -1,6 +1,7 @@
 var Shoppingcart = require('../models/shoppingcart');
 var Pen = require('../models/pens');
 var mongoose = require('mongoose');
+var uniqueid = require('../lib/uniqueid');
 
 var async = require("async");
 
@@ -8,9 +9,6 @@ exports.index = function(req, res){
   
   
     
-  
-  
-  
     function listCart(callback){
         
          if (req.session.auser == "true"){
@@ -98,6 +96,7 @@ exports.add = function(req, res){
          console.log("Такого нету");
          req.session.cart = [];
          req.session.cart[0] = {};
+         req.session.cart[0]._id = uniqueid.uniqueid();
          req.session.cart[0].qty = req.param('qty');
          req.session.cart[0].title = req.param('title');
          req.session.cart[0].code = req.param('code');
@@ -106,6 +105,7 @@ exports.add = function(req, res){
          console.log("Такой есть");
          var counter =  req.session.cart.length;
          req.session.cart[counter] = {};
+         req.session.cart[counter]._id = uniqueid.uniqueid();
          req.session.cart[counter].qty = req.param('qty');
          req.session.cart[counter].code = req.param('code');
          req.session.cart[counter].title = req.param('title');
@@ -122,6 +122,7 @@ exports.add = function(req, res){
          console.log("Такого нету");
          req.session.cart = [];
          req.session.cart[0] = {};
+         req.session.cart[0]._id = uniqueid.uniqueid();
          req.session.cart[0].qty = req.param('qty');
          req.session.cart[0].title = req.param('title');
          req.session.cart[0].code = req.param('code');
@@ -130,6 +131,7 @@ exports.add = function(req, res){
          console.log("Такой есть");
          var counter =  req.session.cart.length;
          req.session.cart[counter] = {};
+         req.session.cart[counter]._id = uniqueid.uniqueid();
          req.session.cart[counter].qty = req.param('qty');
          req.session.cart[counter].code = req.param('code');
          req.session.cart[counter].title = req.param('title');
@@ -205,11 +207,32 @@ exports.remove = function(req, res){
           //var id = mongoose.Types.ObjectId("57a34a23c0ec1d7b2f6984ef");
           //console.log("ObjectId: " + id);
           
-          Shoppingcart.findByIdAndRemove(req.param('id')).exec(
-            function (err, result) {
-             if(err) throw err;
-             callback(null, result);
+          
+          
+          if(req.param('id').length == 32){
+              console.log("This is 32-bit case");
+              for (var i = 0; i < req.session.cart.length; i++) {
+                  if (req.session.cart[i]._id == req.param('id')) {
+                      console.log("Delete"+req.session.cart[i]._id);
+                      req.session.cart.splice(i,1);
+                  }
+              }
+              var result = req.session.cart;
+              callback(null, result);
+          }
+          if(req.param('id').length == 24){
+              console.log("This is 24-bit case");
+                 Shoppingcart.findByIdAndRemove(req.param('id')).exec(
+                  function (err, result) {
+                  if(err) throw err;
+                  callback(null, result);
            });
+          }
+          
+          
+          
+          
+       
            
     }
